requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
future==0.16.0
idna==2.6
logzero==1.5.0
oauthlib==2.0.6
python-twitter==3.4
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
